+++ python/tvm/relax/op/__init__.py
@@ -19,6 +19,7 @@
 
 # Operators
 from .base import *
+from .manipulate import *
 from .op_attrs import *
 from .reduce import *
 from .tensor import *
