+++ python/tvm/script/ir_builder/relax/ir.py
@@ -51,6 +51,7 @@ from tvm.relax.op import (
     multiply,
     negative,
     nn,
+    permute_dims,
     print,
     reshape,
     shape_of,
@@ -61,7 +62,6 @@ from tvm.relax.op import (
     subtract,
     sum,
     tanh,
-    transpose,
     unique,
     variance,
     wrap_param,
@@ -456,6 +456,7 @@ __all__ = [
     "negative",
     "nn",
     "output",
+    "permute_dims",
     "print",
     "reshape",
     "shape_of",
@@ -467,7 +468,6 @@ __all__ = [
     "sum",
     "tanh",
     "tensor",
-    "transpose",
     "unique",
     "variance",
     "wrap_param",
