+++ tests/python/relax/test_tvmscript_parser_transform_ops.py
@@ -115,125 +115,5 @@ def test_wrap_param():
     _check(expected, bb.get()["main"])
 
 
-def test_reshape():
-    @R.function
-    def foo(x: R.Tensor((1, 2, 3, 4), "float32")) -> R.Tensor((8, 3), "float32"):
-        gv: R.Tensor((8, 3), "float32") = R.reshape(x, (8, 3))
-        return gv
-
-    x = relax.Var("x", R.Tensor((1, 2, 3, 4), "float32"))
-    bb = relax.BlockBuilder()
-    with bb.function("foo", [x]):
-        gv = bb.emit(relax.op.reshape(x, new_shape=(8, 3)))
-        bb.emit_func_output(gv)
-
-    _check(foo, bb.get()["foo"])
-
-
-def test_reshape_infer_dim():
-    @R.function
-    def foo(x: R.Tensor((1, 2, 3, 4), "float32")) -> R.Tensor((8, 1, 3), "float32"):
-        gv: R.Tensor((8, 1, 3), "float32") = R.reshape(x, (8, -1, 3))
-        return gv
-
-    x = relax.Var("x", R.Tensor((1, 2, 3, 4), "float32"))
-    bb = relax.BlockBuilder()
-    with bb.function("foo", [x]):
-        gv = bb.emit(relax.op.reshape(x, new_shape=(8, -1, 3)))
-        bb.emit_func_output(gv)
-
-    _check(foo, bb.get()["foo"])
-
-
-def test_transpose():
-    @R.function
-    def foo(x: R.Tensor((1, 2, 3, 4), "float32")) -> R.Tensor((2, 4, 3, 1), "float32"):
-        gv: R.Tensor((2, 4, 3, 1), "float32") = R.transpose(x, axes=[1, -1, 2, -4])
-        return gv
-
-    x = relax.Var("x", R.Tensor((1, 2, 3, 4), "float32"))
-    bb = relax.BlockBuilder()
-    with bb.function("foo", [x]):
-        gv = bb.emit(relax.op.transpose(x, axes=[1, -1, 2, -4]))
-        bb.emit_func_output(gv)
-
-    _check(foo, bb.get()["foo"])
-
-
-def test_transpose_none_arg():
-    @R.function
-    def foo(x: R.Tensor((1, 2, 3, 4), "float32")) -> R.Tensor((4, 3, 2, 1), "float32"):
-        gv: R.Tensor((4, 3, 2, 1), "float32") = R.transpose(x)
-        return gv
-
-    x = relax.Var("x", R.Tensor((1, 2, 3, 4), "float32"))
-    bb = relax.BlockBuilder()
-    with bb.function("foo", [x]):
-        gv = bb.emit(relax.op.transpose(x))
-        bb.emit_func_output(gv)
-
-    _check(foo, bb.get()["foo"])
-
-
-def test_expand_dims():
-    @R.function
-    def foo(x: R.Tensor((2, 3, 4), "float32")) -> R.Tensor((2, 1, 1, 1, 3, 1, 4, 1), "float32"):
-        gv: R.Tensor((2, 1, 1, 1, 3, 1, 4, 1), "float32") = R.expand_dims(x, axis=[-1, 1, -6, 3, 5])
-        return gv
-
-    x = relax.Var("x", R.Tensor((2, 3, 4), "float32"))
-    bb = relax.BlockBuilder()
-    with bb.function("foo", [x]):
-        gv = bb.emit(relax.op.expand_dims(x, axis=[-1, 1, -6, 3, 5]))
-        bb.emit_func_output(gv)
-
-    _check(foo, bb.get()["foo"])
-
-
-def test_squeeze():
-    @R.function
-    def foo(x: R.Tensor((2, 1, 3, 1, 1, 4), "float32")) -> R.Tensor((2, 3, 4), "float32"):
-        gv: R.Tensor((2, 3, 4), "float32") = R.squeeze(x)
-        return gv
-
-    x = relax.Var("x", R.Tensor((2, 1, 3, 1, 1, 4), "float32"))
-    bb = relax.BlockBuilder()
-    with bb.function("foo", [x]):
-        gv = bb.emit(relax.op.squeeze(x))
-        bb.emit_func_output(gv)
-
-    _check(foo, bb.get()["foo"])
-
-
-def test_squeeze_with_indices():
-    @R.function
-    def foo(x: R.Tensor((2, 1, 3, 1, 1, 4), "float32")) -> R.Tensor((2, 3, 1, 4), "float32"):
-        gv: R.Tensor((2, 3, 1, 4), "float32") = R.squeeze(x, axis=[3, -5])
-        return gv
-
-    x = relax.Var("x", R.Tensor((2, 1, 3, 1, 1, 4), "float32"))
-    bb = relax.BlockBuilder()
-    with bb.function("foo", [x]):
-        gv = bb.emit(relax.op.squeeze(x, axis=[3, -5]))
-        bb.emit_func_output(gv)
-
-    _check(foo, bb.get()["foo"])
-
-
-def test_flatten():
-    @R.function
-    def foo(x: R.Tensor((3, 4, 5), "float32")) -> R.Tensor((60,), "float32"):
-        gv: R.Tensor((60,), "float32") = R.flatten(x)
-        return gv
-
-    x = relax.Var("x", R.Tensor((3, 4, 5), "float32"))
-    bb = relax.BlockBuilder()
-    with bb.function("foo", [x]):
-        gv = bb.emit(relax.op.flatten(x))
-        bb.emit_func_output(gv)
-
-    _check(foo, bb.get()["foo"])
-
-
 if __name__ == "__foo__":
     tvm.testing.main()
